third_party/WebKit/LayoutTests/inspector/runtime/runtime-es6-setSymbolPropertyValue.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script>

var object1 = { foo: 1 };
var symbol1 = Symbol("a");
object1[symbol1] = 2;

function dumpSymbolProperty(label)
{
    console.log("===== " + label + " =====");
    console.log(object1[symbol1]);
    console.log("");
}

function test()
{
    var obj1, name;

    InspectorTest.runTestSuite([
        function testSetUp(next)
        {
            InspectorTest.evaluateInPage("dumpSymbolProperty('Initial')", step0);

            async function step0()
            {
                var result = await InspectorTest.RuntimeAgent.evaluate("object1");
                obj1 = InspectorTest.runtimeModel.createRemoteObject(result);
                result = await InspectorTest.RuntimeAgent.evaluate("symbol1");
                name = SDK.RemoteObject.toCallArgument(InspectorTest.runtimeModel.createRemoteObject(result));
                next();
            }
        },

        async function testSetSymbolPropertyValue(next)
        {
            await obj1.setPropertyValue(name, "3");
            InspectorTest.evaluateInPage("dumpSymbolProperty('Set property')", next);
        },

        async function testDeleteSymbolProperty(next)
        {
            await obj1.deleteProperty(name);
            InspectorTest.evaluateInPage("dumpSymbolProperty('Delete property')", next);
        }
    ]);
}

</script>
</head>

<body onload="runTest()">
<p>
Tests editing Symbol properties.
</p>

</body>
</html>
